Fixing sheet-selection display of current selection
-- What
The issue was that selector.selectionFrame is
data-relative. Currently, the pieces have no single method for
determining a Frame corresponding to the actual view cell locations of
a given selection (ie, translating back from data relative origin and
corner to the viewFrame origin and corner, intersecting only with
what's visible).

This functionality needs to be rolled into Selector.js in order to
keep this clean and well organized. Only then can we move styling
stuff out of Selector.js.