Row and Column tabs update with shifting
-- What
Prior to this commit, any present row and/or column tabs were not
updating their labels whenever the view frame shifted. That has now
been resolved.

To implement this, we are also introducing the new
`sheet-view-shifted` event, which is triggered whenever a navigation
on the PrimaryFrame results in a view shift of some kind. Tabs are
updated in the GridSheet's own handler for this event, `handleViewShift()`.

Importantly, <row-tab> and <column-tab> elements are now rendered with
an attribute that says whether or not they appear in a locked
row. This helps us determine that such tabs should *not* be updated
during a shift event handler.