Implementing highlight tracking in tabs of selection/cursor
-- What
Whenever the cursor or selection changes, any row or column tabs that are
present in the selection (or align with the cursor) will now have an
attributed called `highlighted="true"` set. Styling this attribute
allows the sheet to visually "track" where the cursor and/or selection
is moving along each axis.

Default styling is of the tab background colors when this attribute is
set is present in the shadow dom style definition for RowTab and
ColumnTab. You can override the background color using the variable `--tracking-highlight-color`