Switching the display of locked rows/cols to use SelectionElement*
-- What
We introduce two new types of elements: `<locked-rows>` and
`<locked-columns>`, which are subclasses of the
`SelectionElement` (found in `LockedSelectionElement.js`). They are
used for styling regions on the grid associated with any locked
row/column frames that are present, if any.

This obviates some of the styling of individual view cell elements
when they become active in a locked row and/or column frame.