Initial commit of test direct cell editing
-- What
This commit adds features to SheetCell so that it can be directly
edited simply by typing a non-arrow or functional key when it is the
current cursor.

For the moment it is not completely working -- there are some issues
with contenteditable and cross-browser selecting that need to be
worked out (it doesn't work at all in Chrome, for example).

However, the high level design is sound. We dispatch a custom event
called `cell-edited` when Enter is pressed (if shift is not held).