Fixing arrow key handling in editing sheet cell
-- What
Previously, keydown events for arrow key were bubbling up to the
KeyHandler, whose default handlers for these keys is navigation. This
prevented the "normal" behavior of moving the text cursor around in a
cell that is being edited. SheetCell now intercepts arrow keydown
events (when editing is true) and stops the propagation, so that the
user can have the normal expected behavior.