Fixing issue where keyboard copy/paste no longer worked
-- What
KeyHandler was interfering with the binding of Ctrl+c and Ctrl+v for
copying and pasting. Instead, when copying or pasting the cursor cell
would go into editing mode since an alpha key was pressed.

We now check to see if any of the appropriate modifier keys has been
pressed before triggering an edit on a cell.