Fixing behavior of focus in cell input blur; Adding new edit behav.
-- What
This commit takes care of two related issues.

First, it ensures that edited cell contents are saved (and trigger a
corresponding cell-edited event) whenever the active edit cell is
blurred/navigated away from, be that from a tab, enter, or mouse
event. The exception is the Escape key, whose handler caches the
original value and restores it again on blur.

Second, and in the course of writing the first, a bug was found
whereby keys like tab, enter, and escape would stop editing on a sheet
cell (thereby blurring it), but fail to restore focus to their parent
sheet. This has now been resolved with a new helper method.