[Incomplete] Refactoring cell auto resizing
-- Notes
1. I have added a working setTimeout implementation so this thing
isn't triggering constant resizes (ie, a delay before doing the actual
resizing in case more observed resizes happen in the meantime)
2. sheet.cellHeight was not being set in the grid template layout;
instead it was `1fr`. This was *definitely* screwing up our resize
calculations before, as we were expecting the sheet's cellHeight to be
the default value and making those measurments accordingly.

-- TODO
1. Reimplement vertical shrinking, horizontal filling, horizontal
shrinking
2. Fiddle around with the possibility of auto adjusting the container
size to perfectly fit the table, and other methods of making it look a
little better.