Adding IDBDataFrame subclass
-- What
This commit introduces an example subclass of the new async-enabled
DataFrame superclass. It is designed to interact with the browser's
IndexedDB (using the external 'idb' helper module).

This commit also updates the base GridSheet to use the IDBDataFrame as
its dataFrame, so you can test it right away at localhost:8000/examples.