Adding init pattern for DataStore
-- What
This commit adds an asychronous `init()` method to all DataStore
instances. When resolved, the init method will set the DataStore's
`isReady` property to `true`.

GridSheet and PrimaryFrame now make use of init() / isReady, so that
in the future we can asynchronously set up a sheet and have it display
a loader in the meantime.