Adding #setDataStore capability
-- What
This commit adds a setDataStore method and lifecycle handling to the
sheet. Additionally, it introduces a new required lifecycle method on
all DataStore classes called detach, which is a async function that
will be called on outgoing data stores when a sheet is setting a new
one.

This functionality is designed so that we can easily swap in new
DataStore types/classes as needed.

We have included a dummy IdiotDataStore as an example that you can try
in the browser.

Tests included.