Important PrimaryFrame and APSheet interaction changes
-- What
This commit has two main purposes, both dealing with the interaction
of APSheet, PrimaryFrame, and rendering/loading of potentially
asynchronous data from a DataStore.

1. PrimaryFrame now has a `handleDataChanged()` method that will be
called during a normal APSheet onDataChanged invocation, which is
itself the result of a call to `notify()` from within a DataStore.

This means that any data updates that the DataStore notifies the sheet
about can be captured by the PrimaryFrame. The implementation of
handleDataChanged ensures that the PrimaryFrame will do no special
lookups if the changed data is not within any of its data relative
frames (ie, if the user is not seeing those data cells on the screen
at the moment). This minimizes compute and DOM updates.

2. Any call to APSheet's `render()` that produces a _new_ PrimaryFrame
instance -- such as initial render or resizing -- will make a bulk
asynchronous call to the DataStore in order to attempt to retrieve any
persistent data that might currently be in the view.

This type of request takes advantage of a new kind of persistent
DataStore call named `getPersistentRangeAt()` which takes origin and
corner coordinate pairs and attempts to fetch all the data for it at
once. This prevents tons of little requests going through in favor of
several larger ones. The default DataStore implementation is a no-op,
since subclasses that actually persist data should handle how they
retrieve it.