Adding initial scaffolding for contextmenu custom event
-- What
Because our column and row tabs are in the shadow dom, we cannot
handle mouse events on them from outside of the APSheet element.

This precludes us from specifically handling contextmenu events on
those elements from consuming applications -- in particular, a
Preact/React application that renders an ap-sheet custom element.

The solution started in this commit adds a new observed attribute for
turning a general contextmenu eventing mechanism on, which will add
appropriate contextmenu event listeners to the sheet itself, but also
the shadowed row and column tabs. These listeners take a small amount
of information needed and dispatch a custom event called
`sheetcontextmenu` that has information about which elements were
actually clicked.

This is a temporary and hacky solution. The best bet would be to have
more generalized custom events (for example, whenever an ap-sheet
re-renders cell and tab elements), along with having the important
elements out in the live dom so proper eventing can be added to them
by consumers.