Removing sites for redundant fetch calls
-- What
During an audit of network calls made during a persistent DataStore
implementation (see `CSVDataStore` in JLApp), we noticed that there
were many redundant network requests being made at the same time.

Going through the calls sites in the devtools revealed several places
where requests could be called multiple times. This commit removes
those extra call sites, and ensures that for certain selector
movements and changes, a request to get data from the persistent part
of a DataStore should only happen once per update.

Note that a lot of this has to do with the logic already being present
in some part of a sheet's `render()` call -- the extra call sites
where in addition to this in most cases.