Updating baseFrame to be in sync with DataStore max
-- What
With recent changes in the call sites for async DataFrame requests, it
was possible for the sheet's baseFrame extent to become out of sync
with the maximum possible values for a given DataStore.

It's important to recap how computing the max values occurs in the two
types of DataStore we've created for JL:
* plain DataStore -- The max extent value is computed by taking the
maximum x and y values that are currently present in the dictionary's
keys;
* CSVDataStore -- The max extent is retrieved _from the server when
any data request is made_. This means that at the beginning, there is
an arbitrary baseFrame corner set and the DataStore's max values are
nothing.

This commit fixes an issue whereby a user would not be able to scroll
to see the full data, because the baseFrame assumed it was at the end
when it wasn't. Now they should be back in sync.